+++ packages/remix-dev/compiler/loaders.ts
@@ -3,10 +3,13 @@ import type * as esbuild from "esbuild";
 
 export const loaders: { [ext: string]: esbuild.Loader } = {
   ".aac": "file",
+  ".avif": "file",
   ".css": "file",
   ".eot": "file",
   ".flac": "file",
   ".gif": "file",
+  ".gql": "text",
+  ".graphql": "text",
   ".ico": "file",
   ".jpeg": "file",
   ".jpg": "file",
@@ -33,6 +36,7 @@ export const loaders: { [ext: string]: esbuild.Loader } = {
   ".webp": "file",
   ".woff": "file",
   ".woff2": "file",
+  ".zip": "file",
 };
 
 export function getLoaderForFile(file: string): esbuild.Loader {