+++ packages/remix-dev/modules.ts
@@ -2,6 +2,10 @@ declare module "*.aac" {
   let asset: string;
   export default asset;
 }
+declare module "*.avif" {
+  let asset: string;
+  export default asset;
+}
 declare module "*.css" {
   let asset: string;
   export default asset;
@@ -15,6 +19,14 @@ declare module "*.flac" {
   export default asset;
 }
 declare module "*.gif" {
+  let asset: string;
+  export default asset;
+}
+declare module "*.gql" {
+  let asset: string;
+  export default asset;
+}
+declare module "*.graphql" {
   let asset: string;
   export default asset;
 }
@@ -60,6 +72,10 @@ declare module "*.png" {
   let asset: string;
   export default asset;
 }
+declare module "*.sql" {
+  let asset: string;
+  export default asset;
+}
 declare module "*.svg" {
   let asset: string;
   export default asset;
@@ -89,6 +105,10 @@ declare module "*.woff2" {
   export default asset;
 }
 declare module "*.webmanifest" {
+  let asset: string;
+  export default asset;
+}
+declare module "*.zip" {
   let asset: string;
   export default asset;
 }