+++ packages/create-remix/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `create-remix`
 
+## 1.6.8
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.6.8`
+
 ## 1.6.7
 
 ### Patch Changes