+++ packages/remix-vercel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/vercel`
 
+## 1.6.8
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.6.8`
+
 ## 1.6.7
 
 ### Patch Changes